#10982 Side panels should be smaller (#11009)
* ##10982 Side panels should be smaller

* fix failing tests

* requested changes
